+++ src/components/App.js
@@ -11,7 +11,7 @@ import { getError } from 'console/state/auth/selectors';
 import AuthButton from 'console/components/auth/AuthButton';
 import QueryActions from 'console/components/data/QueryActions';
 import QueryAuth0 from 'console/components/data/QueryAuth0';
-import AppRouter from 'console/components/router';
+import AppRouter from 'console/components/AppRouter';
 import CircleLogo from 'console/components/svg/CircleLogo';
 const { Header, Content } = Layout;
